# openapi.yaml
openapi: 3.0.3
info:
  title: Beacon Backend API
  version: 1.1.0
paths:
...
  /me/bootstrap:
    post:
      summary: Create/update current user profile from auth context
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [full_name, role]
              properties:
                full_name:
                  type: string
                phone_number:
                  type: string
                role:
                  type: string
                  enum: [citizen, student]
      responses:
        "200":
          description: Profile upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "400":
          $ref: "#/components/responses/BadRequest"
  /me:
    get:
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: Update current authenticated user profile
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchMeRequest"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "400":
          $ref: "#/components/responses/BadRequest"
  /users:
    get:
      summary: Get current authenticated user (compatibility alias)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
  /users/search:
    get:
      summary: Search app users for friend discovery
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Case-insensitive full-name search query. Minimum length is 2 characters.
          schema:
            type: string
            minLength: 2
      responses:
        "200":
          description: Up to 20 matching public user records with current relationship state
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSearchResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /notifications:
    get:
      summary: List current authenticated user's notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserNotificationDto"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /notifications/{id}/read:
    patch:
      summary: Mark current user's notification as read
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserNotificationReadResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  schemas:
...
    UserDto:
      type: object
      required: [id, firebase_uid, email, full_name, role]
      properties:
        id:
          type: integer
        firebase_uid:
          type: string
        email:
          type: string
          format: email
          nullable: true
        full_name:
          type: string
        phone_number:
          type: string
          nullable: true
        role:
          type: string
          enum: [citizen, student]
        profile_image_url:
          type: string
          format: uri
          nullable: true
        status:
          type: string
          enum: [active, deactivated]
    FriendshipStatus:
      type: string
      enum: [none, already_friends, incoming_pending, outgoing_pending]
    UserSearchResult:
      type: object
      required: [id, full_name, friendship_status]
      properties:
        id:
          type: integer
        full_name:
          type: string
        beacon_code:
          type: string
          nullable: true
        friendship_status:
          $ref: "#/components/schemas/FriendshipStatus"
...
    PatchMeRequest:
      type: object
      properties:
        full_name:
          type: string
          minLength: 2
          maxLength: 50
        email:
          type: string
          format: email
        phone_number:
          type: string
          pattern: "^\\+?\\d{10,20}$"
        role:
          type: string
          enum: [citizen, student]
        profile_image_url:
          type: string
          format: uri
          nullable: true
    UserNotificationType:
      type: string
      enum: [incident_update, sos_update]
    UserNotificationMetadata:
      type: object
      properties:
        incident_id:
          type: integer
          nullable: true
        sos_id:
          type: integer
          nullable: true
        status:
          type: string
          nullable: true
        assigned_unit:
          type: string
          nullable: true
        assigned_department:
          type: string
          nullable: true
        fallback_route:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
    UserNotificationDto:
      type: object
      properties:
        id:
          type: integer
        recipient_user_id:
          type: integer
        type:
          $ref: "#/components/schemas/UserNotificationType"
        title:
          type: string
        message:
          type: string
        metadata:
          $ref: "#/components/schemas/UserNotificationMetadata"
        is_read:
          type: boolean
        created_at:
          type: string
          format: date-time
    UserNotificationReadResponse:
      type: object
      properties:
        message:
          type: string
        notification:
          $ref: "#/components/schemas/UserNotificationDto"
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...